storage: refactoring to use api version 2022-05-01
