+++ internal/services/storage/parse/object_replication_test.go
@@ -5,7 +5,7 @@ package parse
 import (
 	"testing"
 
-	"github.com/hashicorp/go-azure-sdk/resource-manager/storage/2021-04-01/objectreplicationpolicies"
+	"github.com/hashicorp/go-azure-sdk/resource-manager/storage/2022-05-01/objectreplicationpolicies"
 )
 
 func TestObjectReplicationIDFormatter(t *testing.T) {
