+++ internal/services/storage/storage_object_replication_resource_test.go
@@ -141,7 +141,7 @@ func (r StorageObjectReplicationResource) Exists(ctx context.Context, client *cl
 	if err != nil {
 		return nil, err
 	}
-	dstResp, err := client.Storage.ObjectReplicationClient.Get(ctx, id.Dst)
+	dstResp, err := client.Storage.ResourceManager.ObjectReplicationPolicies.Get(ctx, id.Dst)
 	if err != nil {
 		if response.WasNotFound(dstResp.HttpResponse) {
 			return utils.Bool(false), nil
@@ -149,7 +149,7 @@ func (r StorageObjectReplicationResource) Exists(ctx context.Context, client *cl
 		return nil, fmt.Errorf("retrieving %q: %+v", id, err)
 	}
 
-	srcResp, err := client.Storage.ObjectReplicationClient.Get(ctx, id.Src)
+	srcResp, err := client.Storage.ResourceManager.ObjectReplicationPolicies.Get(ctx, id.Src)
 	if err != nil {
 		if response.WasNotFound(srcResp.HttpResponse) {
 			return utils.Bool(false), nil
